+++ relay-server/src/services/buffer/envelope_buffer/mod.rs
@@ -78,50 +78,50 @@ impl PolymorphicEnvelopeBuffer {
 
     /// Adds an envelope to the buffer.
     pub async fn push(&mut self, envelope: Box<Envelope>) -> Result<(), EnvelopeBufferError> {
-        let partition_tag = self.partition_tag().to_owned();
-        match self {
-            Self::Sqlite(buffer) => {
-                relay_statsd::metric!(
-                    distribution(RelayDistributions::BufferEnvelopeBodySize) =
-                        envelope.items().map(Item::len).sum::<usize>() as u64,
-                    partition_id = &partition_tag
-                );
-                relay_statsd::metric!(
-                    timer(RelayTimers::BufferPush),
-                    partition_id = &partition_tag,
-                    { buffer.push(envelope).await }
-                )
-            }
-            Self::InMemory(buffer) => buffer.push(envelope).await,
-        }?;
+        relay_statsd::metric!(
+            distribution(RelayDistributions::BufferEnvelopeBodySize, sample = 0.01) =
+                envelope.items().map(Item::len).sum::<usize>() as u64,
+            partition_id = self.partition_tag()
+        );
 
-        Ok(())
+        relay_statsd::metric!(
+            timer(RelayTimers::BufferPush, sample = 0.01),
+            partition_id = self.partition_tag(),
+            {
+                match self {
+                    Self::Sqlite(buffer) => buffer.push(envelope).await,
+                    Self::InMemory(buffer) => buffer.push(envelope).await,
+                }
+            }
+        )
     }
 
     /// Returns a reference to the next-in-line envelope.
     pub async fn peek(&mut self) -> Result<Peek, EnvelopeBufferError> {
-        match self {
-            Self::Sqlite(buffer) => relay_statsd::metric!(
-                timer(RelayTimers::BufferPeek),
-                partition_id = self.partition_tag(),
-                { buffer.peek().await }
-            ),
-            Self::InMemory(buffer) => buffer.peek().await,
-        }
+        relay_statsd::metric!(
+            timer(RelayTimers::BufferPeek, sample = 0.01),
+            partition_id = self.partition_tag(),
+            {
+                match self {
+                    Self::Sqlite(buffer) => buffer.peek().await,
+                    Self::InMemory(buffer) => buffer.peek().await,
+                }
+            }
+        )
     }
 
     /// Pops the next-in-line envelope.
     pub async fn pop(&mut self) -> Result<Option<Box<Envelope>>, EnvelopeBufferError> {
-        let envelope = match self {
-            Self::Sqlite(buffer) => relay_statsd::metric!(
-                timer(RelayTimers::BufferPop),
-                partition_id = self.partition_tag(),
-                { buffer.pop().await }
-            ),
-            Self::InMemory(buffer) => buffer.pop().await,
-        }?;
-
-        Ok(envelope)
+        relay_statsd::metric!(
+            timer(RelayTimers::BufferPop, sample = 0.01),
+            partition_id = self.partition_tag(),
+            {
+                match self {
+                    Self::Sqlite(buffer) => buffer.pop().await,
+                    Self::InMemory(buffer) => buffer.pop().await,
+                }
+            }
+        )
     }
 
     /// Marks a project as ready or not ready.