+++ relay/src/setup.rs
@@ -115,7 +115,7 @@ pub fn init_metrics(config: &Config) -> Result<()> {
         prefix: config.metrics_prefix(),
         host: &addrs[..],
         default_tags,
-        sample_rate: config.metrics_sample_rate(),
+        sample_rate: config.metrics_sample_rate().into(),
         aggregate: config.metrics_aggregate(),
         allow_high_cardinality_tags: config.metrics_allow_high_cardinality_tags(),
     });